Shrink legacy CLI surface and clarify entrypoints
