+++ simulate_cli.py
@@ -35,6 +35,63 @@ from viz import plot_distribution, plot_paths
 
 LOGGER = logging.getLogger(__name__)
 
+SIMULATION_ARG_DEFAULTS: dict[str, object] = {
+    "journal_file": None,
+    "policy_file": None,
+    "tickers": "AAPL",
+    "days": 252,
+    "scenarios": 1000,
+    "max_paths": 100,
+    "no_plots": False,
+    "dt": 1.0,
+    "block_size": 1,
+    "shock_probability": 0.0,
+    "shock_return": -0.15,
+    "seed": None,
+    "model": "historical",
+    "fundamental_probability": None,
+    "market_price": None,
+    "fundamental_certainty": 100.0,
+    "prob_mean_reversion": 0.2,
+    "prob_daily_volatility": 0.03,
+    "start": None,
+    "end": None,
+    "output": None,
+    "cache_dir": None,
+    "refresh_cache": False,
+    "save_simulations": False,
+    "offline_path": None,
+    "offline_only": False,
+    "allow_local_fallback": True,
+    "show": True,
+    "ai_summary": False,
+    "ai_model": "gpt-4o-mini",
+    "annual_cash_yield": 0.04,
+    "min_expected_return": 0.0,
+    "min_prob_up": 0.5,
+    "portfolio_risk_budget_pct": 0.02,
+    "max_var_95_pct": 0.25,
+    "max_drawdown_q95_pct": None,
+    "target_return_pct": None,
+    "max_loss_pct": None,
+    "min_prob_hit_target": None,
+    "max_prob_breach_loss": None,
+    "capital": None,
+    "allow_fractional_shares": False,
+    "minimal": False,
+    "strict": False,
+    "verbose": False,
+    "details": False,
+}
+
+
+def build_simulation_args(**overrides: object) -> argparse.Namespace:
+    """Return a full simulation namespace with legacy-compatible defaults."""
+
+    values = dict(SIMULATION_ARG_DEFAULTS)
+    values.update(overrides)
+    return argparse.Namespace(**values)
+
 
 def _normalise_tickers(ticker_arg: str) -> list[str]:
     requested = [ticker.strip().upper() for ticker in ticker_arg.split(",") if ticker.strip()]
